pyproject.toml
[tool.poetry]
name = "tiralabra"
version = "0.1.0"
description = ""
authors = ["Maija <maija.saisa@helsinki.fi>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
pygame = "^2.1.2"
pillow = "^9.3.0"


[tool.poetry.group.dev.dependencies]
pylint = "^2.15.5"
pytest = "^7.2.0"
coverage = "^6.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
